requirements.txt
BitstampClient==2.1.0
certifi==2017.7.27.1
chardet==3.0.4
coinbase==2.0.6
cycler==0.10.0
functools32==3.2.3.post2
idna==2.5
matplotlib==2.0.2
numpy==1.13.1
pandas==0.20.3
psycopg2==2.7.3
pycrypto==2.6.1
pyparsing==2.2.0
python-dateutil==2.6.1
pytz==2017.2
requests==2.18.2
scipy==0.19.1
seaborn==0.8
six==1.10.0
subprocess32==3.2.7
urllib3==1.22
